Update to spigot 1.9. Remove 1.6.4-1.7.10 compat.
