--- src/com/cnaude/chairs/core/Utils.java
@@ -1,12 +1,0 @@
-package com.cnaude.chairs.core;
-
-import org.bukkit.Bukkit;
-import org.bukkit.entity.Player;
-
-public class Utils {
-
-	public static Player[] getOnlinePlayers() {
-		return Bukkit.getOnlinePlayers();
-	}
-
-}
