+++ src/com/cnaude/chairs/sitaddons/ChairEffects.java
@@ -11,7 +11,6 @@ import org.bukkit.event.player.PlayerLevelChangeEvent;
 import org.bukkit.event.player.PlayerPickupItemEvent;
 
 import com.cnaude.chairs.core.Chairs;
-import com.cnaude.chairs.core.Utils;
 
 public class ChairEffects {
 
@@ -44,7 +43,7 @@ public class ChairEffects {
 		healTaskID = plugin.getServer().getScheduler().scheduleSyncRepeatingTask(plugin, new Runnable() {
 			@Override
 			public void run() {
-				for (Player p : Utils.getOnlinePlayers()) {
+				for (Player p : Bukkit.getOnlinePlayers()) {
 					if (plugin.getPlayerSitData().isSitting(p)) {
 						if (p.hasPermission("chairs.sit.health")) {
 							double pHealthPcnt = (getPlayerHealth(p)) / getMaxPlayerHealth(p) * 100d;
@@ -90,7 +89,7 @@ public class ChairEffects {
 		pickupTaskID = plugin.getServer().getScheduler().scheduleSyncRepeatingTask(plugin, new Runnable() {
 			@Override
 			public void run() {
-				for (Player p : Utils.getOnlinePlayers()) {
+				for (Player p : Bukkit.getOnlinePlayers()) {
 					if (plugin.getPlayerSitData().isSitting(p)) {
 						for (Entity entity : p.getNearbyEntities(1, 2, 1)) {
 							if (entity instanceof Item) {
