+++ src/com/cnaude/chairs/vehiclearrow/v_1_9_R1/NMSChairsArrow.java
@@ -1,0 +1,23 @@
+package com.cnaude.chairs.vehiclearrow.v_1_9_R1;
+
+import net.minecraft.server.v1_9_R1.EntityTippedArrow;
+
+import org.bukkit.Bukkit;
+import org.bukkit.Location;
+import org.bukkit.craftbukkit.v1_9_R1.CraftServer;
+import org.bukkit.craftbukkit.v1_9_R1.CraftWorld;
+
+public class NMSChairsArrow extends EntityTippedArrow {
+
+	public NMSChairsArrow(CraftWorld cworld, Location location) {
+		super(cworld.getHandle());
+		setPositionRotation(location.getX(), location.getY(), location.getZ(), location.getYaw(), location.getPitch());
+		world.addEntity(this);
+		bukkitEntity = new CraftChairsArrow((CraftServer) Bukkit.getServer(), this);
+	}
+
+	@Override
+	public void m() {
+	}
+
+}
